refactor(cli): extract runtime release gate handler
